I removed all the machining stuff, except output window, print window, and backplot code. This will all be done in python, but more automatically. Also, I changed all the property stuff to use classes instead of callback functions.

+++ src/Shape.h
@@ -17,7 +17,7 @@ protected:
 	CBox m_box;
 	TopoDS_Shape m_shape;
 	wxLongLong m_creation_time;
-	float m_opacity;
+	double m_opacity;
 	bool m_volume_found;
 	double m_volume;
 	gp_Pnt m_centre_of_mass;
